+++ electron/ipc/ffmpeg/filters.ts
@@ -1,0 +1,194 @@
+import type { AudioSyncAdjustment, PauseSegment } from "../types";
+
+export function buildAtempoFilters(tempoRatio: number): string[] {
+	if (!Number.isFinite(tempoRatio) || tempoRatio <= 0) {
+		return [];
+	}
+
+	const filters: string[] = [];
+	let remaining = tempoRatio;
+
+	while (remaining < 0.5) {
+		filters.push("atempo=0.5");
+		remaining /= 0.5;
+	}
+
+	while (remaining > 2) {
+		filters.push("atempo=2.0");
+		remaining /= 2.0;
+	}
+
+	if (Math.abs(remaining - 1) > 0.0005) {
+		filters.push(`atempo=${remaining.toFixed(6)}`);
+	}
+
+	return filters;
+}
+
+export function getAudioSyncAdjustment(
+	videoDuration: number,
+	audioDuration: number,
+): AudioSyncAdjustment {
+	if (
+		!Number.isFinite(videoDuration) ||
+		!Number.isFinite(audioDuration) ||
+		videoDuration <= 0 ||
+		audioDuration <= 0
+	) {
+		return { mode: "none", delayMs: 0, tempoRatio: 1, durationDeltaMs: 0 };
+	}
+
+	const durationDeltaMs = Math.round((videoDuration - audioDuration) * 1000);
+	const absDeltaMs = Math.abs(durationDeltaMs);
+	if (absDeltaMs <= 50) {
+		return { mode: "none", delayMs: 0, tempoRatio: 1, durationDeltaMs };
+	}
+
+	const tempoRatio = Math.max(0.5, Math.min(2, audioDuration / videoDuration));
+	const relativeDelta = absDeltaMs / Math.max(videoDuration * 1000, 1);
+
+	if (relativeDelta <= 0.03 || absDeltaMs <= 1500 || durationDeltaMs < 0) {
+		return { mode: "tempo", delayMs: 0, tempoRatio, durationDeltaMs };
+	}
+
+	return { mode: "delay", delayMs: durationDeltaMs, tempoRatio: 1, durationDeltaMs };
+}
+
+export function appendSyncedAudioFilter(
+	filterParts: string[],
+	inputLabel: string,
+	outputLabel: string,
+	adjustment: AudioSyncAdjustment,
+) {
+	const filters: string[] = [];
+
+	if (adjustment.mode === "delay" && adjustment.delayMs > 0) {
+		filters.push(`adelay=${adjustment.delayMs}|${adjustment.delayMs}`);
+	}
+
+	if (adjustment.mode === "tempo") {
+		filters.push(...buildAtempoFilters(adjustment.tempoRatio));
+	}
+
+	filters.push("aresample=async=1:first_pts=0", "asetpts=PTS-STARTPTS");
+	filterParts.push(`${inputLabel}${filters.join(",")}[${outputLabel}]`);
+}
+
+export function formatFfmpegSeconds(milliseconds: number): string {
+	return (milliseconds / 1000).toFixed(3);
+}
+
+export function normalizePauseSegments(
+	pauseSegments: PauseSegment[] | undefined,
+): PauseSegment[] {
+	if (!Array.isArray(pauseSegments) || pauseSegments.length === 0) {
+		return [];
+	}
+
+	const normalized = pauseSegments
+		.map((segment) => {
+			const startMs = Number(segment?.startMs);
+			const endMs = Number(segment?.endMs);
+
+			if (!Number.isFinite(startMs) || !Number.isFinite(endMs)) {
+				return null;
+			}
+
+			const clampedStart = Math.max(0, Math.round(startMs));
+			const clampedEnd = Math.max(0, Math.round(endMs));
+			if (clampedEnd <= clampedStart) {
+				return null;
+			}
+
+			return { startMs: clampedStart, endMs: clampedEnd };
+		})
+		.filter((segment): segment is PauseSegment => !!segment)
+		.sort((left, right) => left.startMs - right.startMs);
+
+	if (normalized.length <= 1) {
+		return normalized;
+	}
+
+	const merged: PauseSegment[] = [{ ...normalized[0] }];
+
+	for (const segment of normalized.slice(1)) {
+		const previous = merged[merged.length - 1];
+		if (segment.startMs <= previous.endMs) {
+			previous.endMs = Math.max(previous.endMs, segment.endMs);
+		} else {
+			merged.push({ ...segment });
+		}
+	}
+
+	return merged;
+}
+
+export function buildPausedAudioFilter(
+	inputLabel: string,
+	outputLabel: string,
+	pauseSegments: PauseSegment[],
+): string | null {
+	if (pauseSegments.length === 0) {
+		return null;
+	}
+
+	const activeSegments: Array<{ startMs: number; endMs?: number }> = [];
+	let cursorMs = 0;
+
+	for (const pauseSegment of pauseSegments) {
+		if (pauseSegment.startMs > cursorMs) {
+			activeSegments.push({ startMs: cursorMs, endMs: pauseSegment.startMs });
+		}
+		cursorMs = Math.max(cursorMs, pauseSegment.endMs);
+	}
+
+	activeSegments.push({ startMs: cursorMs });
+
+	const filterParts: string[] = [];
+	const segmentLabels: string[] = [];
+
+	activeSegments.forEach((segment, index) => {
+		if (typeof segment.endMs === "number" && segment.endMs <= segment.startMs) {
+			return;
+		}
+
+		const segmentLabel = `${outputLabel}_part${index}`;
+		const trimArgs =
+			typeof segment.endMs === "number"
+				? `start=${formatFfmpegSeconds(segment.startMs)}:end=${formatFfmpegSeconds(segment.endMs)}`
+				: `start=${formatFfmpegSeconds(segment.startMs)}`;
+
+		filterParts.push(`[${inputLabel}]atrim=${trimArgs},asetpts=PTS-STARTPTS[${segmentLabel}]`);
+		segmentLabels.push(`[${segmentLabel}]`);
+	});
+
+	if (segmentLabels.length === 0) {
+		return null;
+	}
+
+	if (segmentLabels.length === 1) {
+		filterParts.push(`${segmentLabels[0]}anull[${outputLabel}]`);
+	} else {
+		filterParts.push(
+			`${segmentLabels.join("")}concat=n=${segmentLabels.length}:v=0:a=1[${outputLabel}]`,
+		);
+	}
+
+	return filterParts.join(";");
+}
+
+export function parseFfmpegDurationSeconds(stderr: string): number | null {
+	const match = stderr.match(/Duration:\s+(\d+):(\d+):(\d+(?:\.\d+)?)/i);
+	if (!match) {
+		return null;
+	}
+
+	const hours = Number(match[1]);
+	const minutes = Number(match[2]);
+	const seconds = Number(match[3]);
+	if (![hours, minutes, seconds].every(Number.isFinite)) {
+		return null;
+	}
+
+	return hours * 3600 + minutes * 60 + seconds;
+}
