+++ electron/ipc/recording/ffmpeg.ts
@@ -1,0 +1,204 @@
+import type { ChildProcessWithoutNullStreams } from "node:child_process";
+import type { SelectedSource } from "../types";
+import {
+	ffmpegCaptureOutputBuffer,
+} from "../state";
+import { getScreen } from "../utils";
+import { resolveWindowsCaptureDisplay } from "../windowsCaptureSelection";
+import { resolveLinuxWindowBounds } from "../cursor/bounds";
+
+export function getDisplayBoundsForSource(source: SelectedSource) {
+	return resolveWindowsCaptureDisplay(
+		source,
+		getScreen().getAllDisplays(),
+		getScreen().getPrimaryDisplay(),
+	).bounds;
+}
+
+export async function buildFfmpegCaptureArgs(source: SelectedSource, outputPath: string) {
+	const commonOutputArgs = [
+		"-an",
+		"-c:v",
+		"libx264",
+		"-preset",
+		"veryfast",
+		"-pix_fmt",
+		"yuv420p",
+		"-movflags",
+		"+faststart",
+		outputPath,
+	];
+
+	if (process.platform === "win32") {
+		if (source?.id?.startsWith("window:")) {
+			const windowTitle =
+				typeof source.windowTitle === "string"
+					? source.windowTitle.trim()
+					: source.name.trim();
+			if (!windowTitle) {
+				throw new Error("Missing window title for FFmpeg window capture");
+			}
+
+			return [
+				"-y",
+				"-f",
+				"gdigrab",
+				"-framerate",
+				"60",
+				"-draw_mouse",
+				"0",
+				"-i",
+				`title=${windowTitle}`,
+				...commonOutputArgs,
+			];
+		}
+
+		return [
+			"-y",
+			"-f",
+			"gdigrab",
+			"-framerate",
+			"60",
+			"-draw_mouse",
+			"0",
+			"-i",
+			"desktop",
+			...commonOutputArgs,
+		];
+	}
+
+	if (process.platform === "linux") {
+		const displayEnv = process.env.DISPLAY || ":0.0";
+		if (source?.id?.startsWith("window:")) {
+			const bounds = await resolveLinuxWindowBounds(source);
+			if (!bounds) {
+				throw new Error("Unable to resolve Linux window bounds for FFmpeg capture");
+			}
+
+			return [
+				"-y",
+				"-f",
+				"x11grab",
+				"-framerate",
+				"60",
+				"-draw_mouse",
+				"0",
+				"-video_size",
+				`${Math.max(2, bounds.width)}x${Math.max(2, bounds.height)}`,
+				"-i",
+				`${displayEnv}+${Math.round(bounds.x)},${Math.round(bounds.y)}`,
+				...commonOutputArgs,
+			];
+		}
+
+		const bounds = getDisplayBoundsForSource(source);
+		return [
+			"-y",
+			"-f",
+			"x11grab",
+			"-framerate",
+			"60",
+			"-draw_mouse",
+			"0",
+			"-video_size",
+			`${Math.max(2, bounds.width)}x${Math.max(2, bounds.height)}`,
+			"-i",
+			`${displayEnv}+${Math.round(bounds.x)},${Math.round(bounds.y)}`,
+			...commonOutputArgs,
+		];
+	}
+
+	if (process.platform === "darwin") {
+		return [
+			"-y",
+			"-f",
+			"avfoundation",
+			"-capture_cursor",
+			"0",
+			"-framerate",
+			"60",
+			"-i",
+			"1:none",
+			...commonOutputArgs,
+		];
+	}
+
+	throw new Error(`FFmpeg capture is not supported on ${process.platform}`);
+}
+
+export function waitForFfmpegCaptureStart(process: ChildProcessWithoutNullStreams) {
+	return new Promise<void>((resolve, reject) => {
+		const onError = (error: Error) => {
+			cleanup();
+			reject(error);
+		};
+
+		const onExit = (code: number | null) => {
+			cleanup();
+			reject(
+				new Error(
+					ffmpegCaptureOutputBuffer.trim() ||
+						`FFmpeg exited before recording started (code ${code ?? "unknown"})`,
+				),
+			);
+		};
+
+		const timer = setTimeout(() => {
+			cleanup();
+			resolve();
+		}, 900);
+
+		const cleanup = () => {
+			clearTimeout(timer);
+			process.off("error", onError);
+			process.off("exit", onExit);
+		};
+
+		process.once("error", onError);
+		process.once("exit", onExit);
+	});
+}
+
+export function waitForFfmpegCaptureStop(process: ChildProcessWithoutNullStreams, outputPath: string) {
+	return new Promise<string>((resolve, reject) => {
+		const onClose = async (code: number | null) => {
+			cleanup();
+
+			try {
+				const { access } = await import("node:fs/promises");
+				await access(outputPath);
+				if (code === 0 || code === null) {
+					resolve(outputPath);
+					return;
+				}
+
+				if (ffmpegCaptureOutputBuffer.includes("Exiting normally")) {
+					resolve(outputPath);
+					return;
+				}
+			} catch {
+				// handled below
+			}
+
+			reject(
+				new Error(
+					ffmpegCaptureOutputBuffer.trim() ||
+						`FFmpeg exited with code ${code ?? "unknown"}`,
+				),
+			);
+		};
+
+		const onError = (error: Error) => {
+			cleanup();
+			reject(error);
+		};
+
+		const cleanup = () => {
+			process.off("close", onClose);
+			process.off("error", onError);
+		};
+
+		process.once("close", onClose);
+		process.once("error", onError);
+	});
+}
