+++ electron/ipc/utils.ts
@@ -1,0 +1,105 @@
+import { createRequire } from "node:module";
+import fs from "node:fs/promises";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import { app } from "electron";
+import { RECORDINGS_DIR } from "../appPaths";
+import { RECORDINGS_SETTINGS_FILE, AUTO_RECORDING_PREFIX } from "./constants";
+import {
+	customRecordingsDir,
+	setCustomRecordingsDir,
+	recordingsDirLoaded,
+	setRecordingsDirLoaded,
+} from "./state";
+
+const nodeRequire = createRequire(import.meta.url);
+
+export function getScreen() {
+	if (!app.isReady()) {
+		throw new Error(
+			"getScreen() called before app is ready. Ensure all screen access happens after app.whenReady().",
+		);
+	}
+	return nodeRequire("electron").screen as typeof import("electron").screen;
+}
+
+export function normalizePath(filePath: string) {
+	return path.resolve(filePath);
+}
+
+export function normalizeVideoSourcePath(videoPath?: string | null): string | null {
+	if (typeof videoPath !== "string") {
+		return null;
+	}
+
+	const trimmed = videoPath.trim();
+	if (!trimmed) {
+		return null;
+	}
+
+	if (/^file:\/\//i.test(trimmed)) {
+		try {
+			return fileURLToPath(trimmed);
+		} catch {
+			// Fall through and keep best-effort string path below.
+		}
+	}
+
+	return trimmed;
+}
+
+export function parseWindowId(sourceId?: string) {
+	if (!sourceId) return null;
+	const match = sourceId.match(/^window:(\d+)/);
+	return match ? Number.parseInt(match[1], 10) : null;
+}
+
+export function getTelemetryPathForVideo(videoPath: string) {
+	return `${videoPath}.cursor.json`;
+}
+
+export function isAutoRecordingPath(filePath: string) {
+	return path.basename(filePath).startsWith(AUTO_RECORDING_PREFIX);
+}
+
+export async function moveFileWithOverwrite(sourcePath: string, destinationPath: string) {
+	await fs.mkdir(path.dirname(destinationPath), { recursive: true });
+	await fs.rm(destinationPath, { force: true });
+
+	try {
+		await fs.rename(sourcePath, destinationPath);
+	} catch (error) {
+		const nodeError = error as NodeJS.ErrnoException;
+		if (nodeError.code !== "EXDEV") {
+			throw error;
+		}
+
+		await fs.copyFile(sourcePath, destinationPath);
+		await fs.unlink(sourcePath);
+	}
+}
+
+async function loadRecordingsDirectorySetting() {
+	if (recordingsDirLoaded) {
+		return;
+	}
+
+	setRecordingsDirLoaded(true);
+
+	try {
+		const content = await fs.readFile(RECORDINGS_SETTINGS_FILE, "utf-8");
+		const parsed = JSON.parse(content) as { recordingsDir?: unknown };
+		if (typeof parsed.recordingsDir === "string" && parsed.recordingsDir.trim()) {
+			setCustomRecordingsDir(path.resolve(parsed.recordingsDir));
+		}
+	} catch {
+		setCustomRecordingsDir(null);
+	}
+}
+
+export async function getRecordingsDir() {
+	await loadRecordingsDirectorySetting();
+	const targetDir = customRecordingsDir ?? RECORDINGS_DIR;
+	await fs.mkdir(targetDir, { recursive: true });
+	return targetDir;
+}
